% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HERVariumR-package.R
\docType{package}
\name{HERVariumR-package}
\alias{HERVariumR}
\alias{HERVariumR-package}
\title{HERVariumR}
\description{
Fast functional annotation and comparison of HERV lists.
}
\author{
\strong{Maintainer}: Tomàs Montserrat-Ayuso \email{tomas.montserrat@cnag.eu}

}
\keyword{internal}
