% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dashboard.R
\name{generate_hervarium_dashboard}
\alias{generate_hervarium_dashboard}
\title{Generate an interactive HERVariumR dashboard}
\usage{
generate_hervarium_dashboard(
  x,
  output_dir = NULL,
  report_file = NULL,
  title = NULL,
  top_n = 25,
  self_contained = TRUE
)
}
\arguments{
\item{x}{A HERVarium_annotation or HERVarium_comparison object.}

\item{output_dir}{Output directory.}

\item{report_file}{Name of the HTML report.}

\item{title}{Report title.}

\item{top_n}{Number of top features to show in dashboard plots.}

\item{self_contained}{Whether to produce a self-contained HTML file.}
}
\value{
Invisibly returns the path to the generated HTML dashboard.
}
\description{
Generates an HTML dashboard from either a HERVarium_annotation or
HERVarium_comparison object.
}
