% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_files.R
\name{hervarium_file}
\alias{hervarium_file}
\title{Get path to a bundled HERVariumR annotation file}
\usage{
hervarium_file(filename)
}
\arguments{
\item{filename}{Name of a file inside the package \code{extdata} directory.}
}
\value{
Full path to the requested bundled file.
}
\description{
Returns the installed path to a file bundled inside \code{inst/extdata}.
}
